archive/pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"


[project]
name = "sftk"
version = "0.1.0"
description = "Spyfish Aotearoa Toolkit for analyzing marine data"
authors = [
  { name="Wildlife.ai", email="contact@wildlife.ai" }
]
readme = "README.md"
requires-python = ">=3.10"


dependencies = [
  "boto3==1.37.17",
  "botocore>=1.37.17",
  "exceptiongroup==1.2.2",
  "iniconfig==2.1.0",
  "jmespath==1.0.1",
  "numpy==2.2.4",
  "openpyxl>=3.1.5",
  "packaging==24.2",
  "pandas==2.2.3",
  "pluggy==1.5.0",
  "python-dateutil==2.9.0.post0",
  "python-dotenv==1.0.1",
  "pytz==2025.1",
  "requests>=2.32.3",
  "s3transfer==0.11.4",
  "six==1.17.0",
  "streamlit==1.50.0",
  "tomli==2.2.1",
  "tqdm==4.67.1",
  "tzdata==2025.1",
  "urllib3==2.3.0"
]


[project.optional-dependencies]
dev = [
  "pytest==8.3.5",
  "pre_commit==4.2.0"
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]


[tool.setuptools.packages.find]
include = ["sftk*"]

[tool.isort]
profile = "black"

[project.urls]
Homepage = "https://github.com/wildlifeai/Spyfish-Aotearoa-toolkit"
